CookingMinutes changed from Time to decimal

+++ Recipies/Recipies.Api/Models/ExposedObjects/ExposedRecipy.cs
@@ -1,7 +1,5 @@
 ﻿using System;
-using System.Collections.Generic;
 using System.Linq;
-using System.Web;
 
 namespace Recipies.Api.Models.ExposedObjects
 {
@@ -10,7 +8,7 @@ namespace Recipies.Api.Models.ExposedObjects
         public string Name { get; set; }
         public int Rating { get; set; }
         public string ImagesFolder { get; set; }
-        public int CookingMinutes { get; set; }
+        public decimal CookingMinutes { get; set; }
         public string CreatedBy { get; set; }
     }
 }
