+++ Recipies/Recipies.Api/Models/ExposedObjects/ExposedStep.cs
@@ -7,7 +7,7 @@ namespace Recipies.Api.Models.ExposedObjects
     public class ExposedStep
     {
         public int OrderOfPrecedence { get; set; }
-        public int PreparationTime { get; set; }
+        public decimal PreparationTime { get; set; }
         public string Description { get; set; }
     }
 }
